requirements.txt
instapy-cli
praw
pushbullet
pillow